fix(naming): Renamed test files to kebob-case
