+++ test/unit/crypto.js
@@ -1,4 +1,4 @@
-const fixtures = require("./fixtures/Crypto.json")
+const fixtures = require("./fixtures/crypto.json")
 const assert = require("assert")
 const BITBOXSDK = require("../../src/BITBOX")
 const BITBOX = new BITBOXSDK()
@@ -7,9 +7,7 @@ const Buffer = require("safe-buffer").Buffer
 describe("#Crypto", () => {
   describe("#sha256", () => {
     fixtures.sha256.forEach(fixture => {
-      it(`should create SHA256Hash hex encoded ${fixture.hash} from ${
-        fixture.hex
-      }`, () => {
+      it(`should create SHA256Hash hex encoded ${fixture.hash} from ${fixture.hex}`, () => {
         const data = Buffer.from(fixture.hex, "hex")
         const sha256Hash = BITBOX.Crypto.sha256(data).toString("hex")
         assert.equal(sha256Hash, fixture.hash)
@@ -25,9 +23,7 @@ describe("#Crypto", () => {
 
   describe("#ripemd160", () => {
     fixtures.ripemd160.forEach(fixture => {
-      it(`should create RIPEMD160Hash hex encoded ${fixture.hash} from ${
-        fixture.hex
-      }`, () => {
+      it(`should create RIPEMD160Hash hex encoded ${fixture.hash} from ${fixture.hex}`, () => {
         const data = Buffer.from(fixture.hex, "hex")
         const ripemd160 = BITBOX.Crypto.ripemd160(data).toString("hex")
         assert.equal(ripemd160, fixture.hash)
@@ -43,9 +39,7 @@ describe("#Crypto", () => {
 
   describe("#hash256", () => {
     fixtures.hash256.forEach(fixture => {
-      it(`should create double SHA256 Hash hex encoded ${fixture.hash} from ${
-        fixture.hex
-      }`, () => {
+      it(`should create double SHA256 Hash hex encoded ${fixture.hash} from ${fixture.hex}`, () => {
         const data = Buffer.from(fixture.hex, "hex")
         const hash256 = BITBOX.Crypto.hash256(data).toString("hex")
         assert.equal(hash256, fixture.hash)
@@ -61,9 +55,7 @@ describe("#Crypto", () => {
 
   describe("#hash160", () => {
     fixtures.hash160.forEach(fixture => {
-      it(`should create RIPEMD160(SHA256()) hex encoded ${fixture.hash} from ${
-        fixture.hex
-      }`, () => {
+      it(`should create RIPEMD160(SHA256()) hex encoded ${fixture.hash} from ${fixture.hex}`, () => {
         const data = Buffer.from(fixture.hex, "hex")
         const hash160 = BITBOX.Crypto.hash160(data).toString("hex")
         assert.equal(hash160, fixture.hash)
