+++ test/unit/fixtures/bitcoincash.json
