+++ test/unit/fixtures/mnemonic.json
